Re-route post
