+++ src/Controller/Admin/PostController.php
@@ -4,18 +4,68 @@ namespace App\Controller\Admin;
 
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
 use App\Repository\PostRepository;
+use App\Entity\Post;
+use App\Form\PostType;
 
 #[Route('/admin/post')]
 class PostController extends AbstractController
 {
-    #[Route('/', name: 'app_admin_post')]
+    #[Route('/', name: 'app_admin_post_index')]
     public function index(PostRepository $postRepository): Response
     {
         return $this->render('admin/post/index.html.twig', [
             'posts' => $postRepository->findAll(), 
             'controller_name' => 'PostController'
         ]);
+    }
+
+    #[Route('/new', name: 'app_post_new', methods: ['GET', 'POST'])]
+    public function new(Request $request, PostRepository $postRepository): Response
+    {
+        $post = new Post();
+        $form = $this->createForm(PostType::class, $post);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $postRepository->save($post, true);
+
+            return $this->redirectToRoute('app_post_index', [], Response::HTTP_SEE_OTHER);
+        }
+
+        return $this->renderForm('admin/post/new.html.twig', [
+            'post' => $post,
+            'form' => $form,
+        ]);
+    }
+
+    #[Route('/{id}/edit', name: 'app_post_edit', methods: ['GET', 'POST'])]
+    public function edit(Request $request, Post $post, PostRepository $postRepository): Response
+    {
+        $form = $this->createForm(PostType::class, $post);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $postRepository->save($post, true);
+
+            return $this->redirectToRoute('app_post_index', [], Response::HTTP_SEE_OTHER);
+        }
+
+        return $this->renderForm('admin/post/edit.html.twig', [
+            'post' => $post,
+            'form' => $form,
+        ]);
+    }
+
+    #[Route('/{id}', name: 'app_post_delete', methods: ['POST'])]
+    public function delete(Request $request, Post $post, PostRepository $postRepository): Response
+    {
+        if ($this->isCsrfTokenValid('delete'.$post->getId(), $request->request->get('_token'))) {
+            $postRepository->remove($post, true);
+        }
+
+        return $this->redirectToRoute('app_post_index', [], Response::HTTP_SEE_OTHER);
     }
 }
