+++ templates/admin/admin/index.html.twig
@@ -10,7 +10,7 @@
                 <ul>
                     <li><a href="{{ path('app_admin_user_index') }}">Users</a></li>
                     <li><a href="{{ path('app_admin_category') }}">Categories</a></li>
-                    <li><a href="{{ path('app_admin_post') }}">Posts</a></li>
+                    <li><a href="{{ path('app_admin_post_index') }}">Posts</a></li>
                     <li><a href="{{ path('app_admin_comment') }}">Comments</a></li>
                 </ul>
             </aside>
