+++ templates/admin/post/edit.html.twig
@@ -1,0 +1,13 @@
+{% extends 'base.html.twig' %}
+
+{% block title %}Edit Post{% endblock %}
+
+{% block body %}
+    <h1>Edit Post</h1>
+
+    {{ include('admin/post/_form.html.twig', {'button_label': 'Update'}) }}
+
+    <a href="{{ path('app_admin_post_index') }}">back to list</a>
+
+    {{ include('admin/post/_delete_form.html.twig') }}
+{% endblock %}
