--- templates/post/edit.html.twig
@@ -1,13 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Edit Post{% endblock %}
-
-{% block body %}
-    <h1>Edit Post</h1>
-
-    {{ include('post/_form.html.twig', {'button_label': 'Update'}) }}
-
-    <a href="{{ path('app_post_index') }}">back to list</a>
-
-    {{ include('post/_delete_form.html.twig') }}
-{% endblock %}
