+++ templates/post/show.html.twig
@@ -31,10 +31,7 @@
             {{ include('comment/_form.html.twig') }}
         </div>
     </section>
-
-    <a href="{{ path('app_post_index') }}">back to list</a>
-    <a href="{{ path('app_post_edit', {'id': post.id}) }}">edit</a>
-    {{ include('post/_delete_form.html.twig') }}
+    
 </main>
 {% endblock %}
